Проект создан

+++ controllers/cards.js
@@ -1,56 +1,58 @@
 const Card = require('../models/card');
+const { BadRequestError } = require('../errors/BadRequestError')
+const { ForbiddenError } = require('../errors/ForbiddenError')
+const { NotFoundError } = require('../errors/NotFoundError')
 
-exports.getCards = async (req, res) => {
+const CODE_OK_200 = 200;
+const CODE_OK_201 = 201;
+
+exports.getCards = async (req, res, next) => {
   try {
-    const CODE_OK_200 = 200;
     const cards = await Card.find({});
     return res.status(CODE_OK_200).send(cards);
   } catch (err) {
-    const ERROR_CODE_500 = 500;
-    return res.status(ERROR_CODE_500).send({ message: 'Ошибка по умолчанию.' });
+    next(err);
   }
 };
 
-exports.createCard = async (req, res) => {
+exports.createCard = async (req, res, next) => {
   try {
-    const CODE_OK_201 = 201;
     const { name, link } = req.body;
     const сard = new Card({ name, link, owner: req.user._id });
     return res.status(CODE_OK_201).send(await сard.save());
   } catch (err) {
-    const ERROR_CODE_400 = 400;
-    const ERROR_CODE_500 = 500;
     if (err.name === 'ValidationError') {
-      return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные при создании карточки.' });
+      next(new BadRequestError('Переданы некорректные данные при создании карточки.'));
+    }else{
+      next(err);
     }
-    return res.status(ERROR_CODE_500).send({ message: 'Ошибка по умолчанию.' });
   }
 };
 
-exports.deleteCard = async (req, res) => {
+exports.deleteCard = async (req, res, next) => {
   try {
-    const CODE_OK_200 = 200;
-    const ERROR_CODE_404 = 404;
     const deletedCard = await Card.findById(req.params.cardId);
     if (deletedCard) {
-      await Card.findByIdAndRemove(req.params.cardId);
-      return res.status(CODE_OK_200).send({ deletedCard });
+      if (req.user._id === deletedCard.owner._id.toString()) {
+        await Card.findByIdAndRemove(req.params.cardId);
+        res.status(CODE_OK_200).send({ deletedCard });
+      }else{
+        throw new ForbiddenError('Недостаочно прав для удаления карточки' );
+      }
+    }else{
+      throw new NotFoundError('Карточка с указанным _id не найдена.' );
     }
-    return res.status(ERROR_CODE_404).send({ message: 'Карточка с указанным _id не найдена.' });
   } catch (err) {
-    const ERROR_CODE_400 = 400;
-    const ERROR_CODE_500 = 500;
     if (err.name === 'CastError') {
-      return res.status(ERROR_CODE_400).send({ message: 'Карточка с указанным _id не найдена.' });
+      next(new BadRequestError('Карточка с указанным _id не найдена.' ));
+    }else{
+      next(err);
     }
-    return res.status(ERROR_CODE_500).send({ message: 'Ошибка по умолчанию.' });
   }
 };
 
-exports.likeCard = async (req, res) => {
+exports.likeCard = async (req, res, next) => {
   try {
-    const CODE_OK_200 = 200;
-    const ERROR_CODE_404 = 404;
     const likedCard = await Card.findById(req.params.cardId);
     if (likedCard) {
       await Card.findByIdAndUpdate(
@@ -59,22 +61,20 @@ exports.likeCard = async (req, res) => {
         { new: true },
       );
       return res.status(CODE_OK_200).send(likedCard);
+    }else{
+      throw new ForbiddenError('Передан несуществующий _id карточки.' );
     }
-    return res.status(ERROR_CODE_404).send({ message: 'Передан несуществующий _id карточки.' });
   } catch (err) {
-    const ERROR_CODE_400 = 400;
-    const ERROR_CODE_500 = 500;
     if (err.name === 'CastError') {
-      return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные для постановки лайка.' });
+      next(new BadRequestError('Переданы некорректные данные для постановки лайка.'));
+    }else{
+      next(err);
     }
-    return res.status(ERROR_CODE_500).send({ message: 'Ошибка по умолчанию.' });
   }
 };
 
-exports.dislikeCard = async (req, res) => {
+exports.dislikeCard = async (req, res, next) => {
   try {
-    const CODE_OK_200 = 200;
-    const ERROR_CODE_404 = 404;
     const dislikedCard = await Card.findById(req.params.cardId);
     if (dislikedCard) {
       await Card.findByIdAndUpdate(
@@ -83,15 +83,14 @@ exports.dislikeCard = async (req, res) => {
         { new: true },
       );
       return res.status(CODE_OK_200).send(dislikedCard);
+    }else{
+      throw new ForbiddenError('Передан несуществующий _id карточки.' );
     }
-
-    return res.status(ERROR_CODE_404).send({ message: 'Передан несуществующий _id карточки.' });
   } catch (err) {
-    const ERROR_CODE_400 = 400;
-    const ERROR_CODE_500 = 500;
     if (err.name === 'CastError') {
-      return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные для снятии лайка.' });
+      next(new BadRequestError('Переданы некорректные данные для снятии лайка.' ));
+    }else{
+      next(err);
     }
-    return res.status(ERROR_CODE_500).send({ message: 'Ошибка по умолчанию.' });
   }
 };
